+++ test/Tester/StreamingTests/DeactivationTestRunner.cs
@@ -5,6 +5,7 @@ using Orleans;
 using Orleans.Runtime;
 using Orleans.Streams;
 using Orleans.TestingHost;
+using Orleans.TestingHost.Utils;
 using UnitTests.GrainInterfaces;
 
 namespace UnitTests.StreamingTests
