+++ test/Tester/StreamingTests/ImplicitSubscritionRecoverableStreamTestRunner.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Orleans;
 using Orleans.TestingHost;
+using Orleans.TestingHost.Utils;
 using TestGrainInterfaces;
 using UnitTests.Grains;
