+++ test/Tester/StreamingTests/StreamGeneratorProviderTests.cs
@@ -10,6 +10,7 @@ using Orleans;
 using Orleans.Providers.Streams.Generator;
 using Orleans.Streams;
 using Orleans.TestingHost;
+using Orleans.TestingHost.Utils;
 using TestGrainInterfaces;
 using TestGrains;
 using Tester;
