+++ test/TesterInternal/General/Identifiertests.cs
@@ -7,6 +7,7 @@ using Orleans.Runtime;
 using Orleans.Runtime.Configuration;
 using Orleans.Serialization;
 using Orleans.TestingHost;
+using Orleans.TestingHost.Utils;
 using Xunit;
 using Xunit.Abstractions;
