+++ test/TesterInternal/StreamingTests/SingleStreamTestRunner.cs
@@ -5,6 +5,7 @@ using Orleans;
 using Orleans.Runtime;
 using Orleans.Streams;
 using Orleans.TestingHost;
+using Orleans.TestingHost.Utils;
 using Tester;
 using UnitTests.Grains;
 using UnitTests.Tester;
